feat(config-types): Move oracle's types to dedicated file
Also rename `index` to `types` and export all from index
